+++ src/petsys_util/convert_raw_to_singles.cpp
@@ -40,7 +40,7 @@ private:
 	long long 	brStepEnd;
 
 	long long	brTime;
-	unsigned short	brChannelID;
+	unsigned int	brChannelID;
 	float		brToT;
 	float		brEnergy;
 	unsigned short	brTacID;
